Fix TypeScript errors and linting issues
- Fixed icon component usage in all pages (ai-3d-model-generator, ai-email-automation, ai-project-management-pro, ai-services, it-services)
- Created missing AnalyticsContext with proper TypeScript types
- Fixed Jest setup configuration and moved to jest.setup.js
- Fixed unused import warnings
- Fixed RequestInit type issue in apiClient
- All TypeScript errors resolved, build passes successfully
- Linting passes with only minor warnings (no errors)
