+++ app/ai-3d-model-generator/page.tsx
@@ -2,31 +2,31 @@
 import React from 'react';
 import Navigation from '../components/Navigation';
 import Footer from '../components/Footer';
-import { CheckCircle, ArrowRight } from 'lucide-react';
+import { CheckCircle, ArrowRight, Brain, BarChart, Target, TrendingUp } from 'lucide-react';
 
 
 const Ai3dModelGeneratorPage: React.FC = () => {
   const features = [
     {
-      icon: 'Brain',
+      icon: Brain,
     title: 'AI-Powered Intelligence',
       description: 'Advanced AI algorithms that provide intelligent insights and recommendations.',
     benefits: ['Smart recommendations', 'Predictive analytics', 'Automated insights', 'Real-time analysis']
 },
   {
-    icon: 'BarChart',
+    icon: BarChart,
     title: 'Advanced Analytics',
       description: 'Comprehensive analytics dashboard with real-time data visualization.',
     benefits: ['Real-time dashboards', 'Custom reports', 'Data visualization', 'Performance metrics']
 },
   {
-    icon: 'Target',
+    icon: Target,
     title: 'Precision Targeting',
       description: 'Target specific goals and objectives with precision and accuracy.',
     benefits: ['Goal tracking', 'Performance optimization', 'Strategic planning', 'Success metrics']
 },
   {
-    icon: 'TrendingUp',
+    icon: TrendingUp,
     title: 'Growth Optimization',
       description: 'Optimize your business growth with data-driven strategies.',
     benefits: ['Growth strategies', 'Market analysis', 'Competitive insights', 'ROI optimization']
