+++ app/ai-email-automation/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ArrowRight, Mail, CheckCircle, Star } from 'lucide-react';
+import { ArrowRight, Mail, CheckCircle, Star, Brain, BarChart, Zap, Send, Shield, DollarSign, Clock, TrendingUp } from 'lucide-react';
 
 
 import Link from 'next/link';
@@ -21,37 +21,37 @@ export const metadata: Metadata = {
 const AIEmailAutomationPage: React.FC = () => {
   const features = [
     {
-      icon: 'Brain',
+      icon: Brain,
     title: 'AI Email Generation',
       description: 'Generate personalized, high-converting emails using AI that understands your audience and brand voice.',
       benefits: ['Personalized content', 'A/B testing', 'Brand voice training', 'Multi-language support']
 },
   {
-    icon: 'Send',
+    icon: Send,
     title: 'Smart Automation',
       description: 'Create complex email workflows with AI-powered triggers, segmentation, and behavioral targeting.',
       benefits: ['Behavioral triggers', 'Smart segmentation', 'Dynamic content', 'Cross-channel sync']
 },
   {
-    icon: 'TrendingUp',
+    icon: TrendingUp,
     title: 'Performance Analytics',
       description: 'Advanced analytics with AI insights to optimize email performance and improve ROI.',
     benefits: ['Open rate optimization', 'Click-through analysis', 'Revenue tracking', 'Predictive insights']
 },
   {
-    icon: 'Zap',
+    icon: Zap,
     title: 'Automated Sequences',
       description: 'AI-powered email sequences that adapt based on recipient behavior and engagement patterns.',
     benefits: ['Adaptive sequences', 'Behavioral triggers', 'Personalization', 'Optimal timing']
 },
   {
-    icon: 'BarChart3',
+    icon: BarChart,
     title: 'Lead Scoring',
       description: 'AI automatically scores leads based on email engagement and behavior to prioritize follow-ups.',
     benefits: ['Lead scoring', 'Engagement tracking', 'Priority alerts', 'CRM integration']
 },
   {
-    icon: 'Shield',
+    icon: Shield,
     title: 'Deliverability Optimization',
       description: 'AI ensures maximum deliverability with reputation monitoring and content optimization.',
     benefits: ['Reputation monitoring', 'Content optimization', 'Spam prevention', 'Compliance checks']
@@ -126,10 +126,10 @@ const AIEmailAutomationPage: React.FC = () => {
     avatar: 'LC'}
   ]
   const stats = [
-    { number: '300%', label: 'Revenue Increase', icon: 'DollarSign' },
-    { number: '150%', label: 'Conversion Boost', icon: 'TrendingUp' },
-    { number: '20hrs', label: 'Time Saved/Week', icon: 'Clock' },
-    { number: '95%', label: 'Deliverability Rate', icon: 'Shield' }
+    { number: '300%', label: 'Revenue Increase', icon: DollarSign },
+    { number: '150%', label: 'Conversion Boost', icon: TrendingUp },
+    { number: '20hrs', label: 'Time Saved/Week', icon: Clock },
+    { number: '95%', label: 'Deliverability Rate', icon: Shield }
   ]
   return (
     <>
