+++ app/it-services/page.tsx
@@ -2,11 +2,11 @@
 
 
 import React from "react";
-import { CheckCircle, ArrowRight } from "lucide-react";
+import { CheckCircle, ArrowRight, Cloud, Shield, Server, Database } from "lucide-react";
 const ITServicesPage: React.FC = () => {
   const services = [
   {
-      icon: 'Cloud',
+      icon: Cloud,
     title: "Cloud Solutions",
       description: "Scalable cloud infrastructure for modern businesses",
     features: [
@@ -18,7 +18,7 @@ const ITServicesPage: React.FC = () => {
       price: "Starting at $2,000"
 },
   {
-    icon: 'Shield',
+    icon: Shield,
     title: "Cybersecurity",
       description: "Protect your business from evolving threats",
     features: [
@@ -30,7 +30,7 @@ const ITServicesPage: React.FC = () => {
       price: "Starting at $3,500"
 },
   {
-    icon: 'Server',
+    icon: Server,
     title: "DevOps Services",
       description: "Streamline development and deployment processes",
     features: [
@@ -42,7 +42,7 @@ const ITServicesPage: React.FC = () => {
       price: "Starting at $4,000"
 },
   {
-    icon: 'Database',
+    icon: Database,
     title: "Data Analytics",
       description: "Turn your data into actionable insights",
     features: [
